requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
dateparser==1.2.1
distro==1.9.0
exceptiongroup==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fuzzywuzzy==0.18.0
GoogleNews==1.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.10.0
joblib==1.5.1
Levenshtein==0.27.1
lxml==5.4.0
lxml_html_clean==0.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
openai==1.84.0
pandas 
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil 
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.6.1
scipy==1.13.1
sgmllib3k==1.0.0
six
sniffio==1.3.1
soupsieve==2.7
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
